Introduce LanguageCommander
