+++ kifi-backend/modules/search/app/com/keepit/controllers/search/SearchController.scala
@@ -28,6 +28,7 @@ class SearchController @Inject() (
     userSearchFilterFactory: UserSearchFilterFactory,
     searchCommander: SearchCommander,
     augmentationCommander: AugmentationCommander,
+    languageCommander: LanguageCommander,
     userExperimentCommander: RemoteUserExperimentCommander) extends SearchServiceController {
 
   def distSearch() = Action(parse.tolerantJson) { request =>
@@ -112,7 +113,7 @@ class SearchController @Inject() (
     Ok(Json.toJson(searchCommander.distLangFreqs(shards, userId).map { case (lang, freq) => lang.lang -> freq }))
   }
 
-  def distLangFreqs2() = Action(parse.tolerantJson) { request =>
+  def distLangFreqs2() = Action.async(parse.tolerantJson) { request =>
     val json = request.body
     val shardSpec = (json \ "shards").as[String]
     val searchRequest = (json \ "request")
@@ -124,7 +125,9 @@ class SearchController @Inject() (
       case _ => LibraryContext.None
     }
     val shards = (new ShardSpecParser).parse[NormalizedURI](shardSpec)
-    Ok(Json.toJson(searchCommander.distLangFreqs2(shards, userId, libraryContext).map { case (lang, freq) => lang.lang -> freq }))
+    languageCommander.distLangFreqs2(shards, userId, libraryContext).map { freqs =>
+      Ok(Json.toJson(freqs.map { case (lang, freq) => lang.lang -> freq }))
+    }
   }
 
   def distAugmentation() = Action.async(parse.tolerantJson) { request =>
